+++ APIV2.Mark.Business/Services/CustomerService.cs
@@ -74,7 +74,7 @@ namespace APIV2.Mark.Business.Services
             }
         }
 
-        public ApiResponse<bool> Delete(int id)
+        public ApiResponse<bool> Delete(long id)
         {
             var result = new ApiResponse<bool>();
             try
