+++ APIV2.Mark.Business/Services/WarehouseService.cs
@@ -94,7 +94,7 @@ namespace APIV2.Mark.Business.Services
             }
         }
 
-        public ApiResponse<bool> Delete(int id)
+        public ApiResponse<bool> Delete(long id)
         {
             var result = new ApiResponse<bool>();
             try
