+++ APIV2.Mark.Entities/Dtos/BankDto.cs
@@ -2,7 +2,7 @@
 {
     public class BankDto
     {
-        public int Id { get; set; }
+        public long Id { get; set; }
         public string? NameEn { get; set; }  
         public string? NameAr { get; set; } 
         public string? Address { get; set; }
@@ -10,7 +10,7 @@
         public string? AccountNumber { get; set; }
         public double? Balance { get; set; }
         public string? Code { get; set; }
-        public int? CurrencyId { get; set; }
+        public long? CurrencyId { get; set; }
         public int StatusId { get; set; }
         public DateTime CreationDate { get; set; }
         public long EmpId { get; set; }
