+++ APIV2.Mark.Entities/Dtos/JournalDetailDto.cs
@@ -8,9 +8,9 @@ namespace APIV2.Mark.Entities.Dtos
 {
     public class JournalDetailDto
     {
-        public int Id { get; set; }
-        public int? JournalId { get; set; }
-        public int? AccountId { get; set; }
+        public long Id { get; set; }
+        public long? JournalId { get; set; }
+        public long? AccountId { get; set; }
         public int? SubAccountId { get; set; }
         public double? Debit { get; set; }
         public double? Credit { get; set; }
