+++ APIV2.Mark.Entities/Dtos/OrderDto.cs
@@ -3,7 +3,7 @@ namespace APIV2.Mark.Entities.Dtos
 {
     public class OrderDto
     {
-        public int Id { get; set; }
+        public long Id { get; set; }
         public string OrderNumber { get; set; } = null!;
         public string? Uuid { get; set; }
         public int? OrderState { get; set; }
@@ -17,8 +17,8 @@ namespace APIV2.Mark.Entities.Dtos
         public double? Discount { get; set; }
         public int? PaymentMethod { get; set; }
         public string? InvoiceNumber { get; set; }
-        public int CustomerId { get; set; }
-        public int WarehouseId { get; set; }
+        public long CustomerId { get; set; }
+        public long WarehouseId { get; set; }
         public double? TotalVat { get; set; }
         public long? AccountId { get; set; }
