+++ APIV2.Mark.Entities/Dtos/TransactionOperationsDto.cs
@@ -13,7 +13,7 @@ namespace APIV2.Mark.Entities.Dtos
         public long OperationId { get; set; }
         public DateTime CreationDate { get; set; }
         public string? Description { get; set; }
-        public int EmpId { get; set; }
+        public long EmpId { get; set; }
         public string OperationCode { get; set; } = null!;
     }
 }
